+++ Source/Particles/CMakeLists.txt
@@ -20,6 +20,6 @@ add_subdirectory(Deposition)
 add_subdirectory(ElementaryProcess)
 add_subdirectory(Gather)
 add_subdirectory(ParticleCreation)
-#add_subdirectory(Pusher)
+add_subdirectory(Pusher)
 add_subdirectory(Resampling)
 add_subdirectory(Sorting)
