+++ Source/Particles/Pusher/CMakeLists.txt
@@ -1,0 +1,7 @@
+foreach(D IN LISTS WarpX_DIMS)
+    warpx_set_suffix_dims(SD ${D})
+    target_sources(lib_${SD}
+      PRIVATE
+        ImplicitPushPX.cpp
+    )
+endforeach()
