+++ Source/Particles/Pusher/Make.package
@@ -1,1 +1,3 @@
+CEXE_sources += ImplicitPushPX.cpp
+
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Particles/Pusher
